QEMU-Virt: Implement system shutdown
In order for QEMU to be compatible with Spike, it implements
a simple protocol used for reporting back the simulation's
status, through the memory-mapped "test finisher" device. We
use that protocol to make QEMU exit on system shutdown.

In case the above fails, we fall back to sending an
IPI_EVENT_HALT to every hart on the system.