lib: Don't access address of packed struct member
Taking the address of a packed member may result in an unalligned
pointer access. This can cause slowdowns for RISC-V and breaks
compilation with GCC9.

To fix the problem let's just store the address into an alligned temp
variable before we access it. This seems the simplest solution as we
don't want to unpack the struct.

This fixes https://github.com/riscv/opensbi/issues/81.

Signed-off-by: Alistair Francis <alistair.francis@wdc.com>