lib: Use a circular queue to keep track of sfence related IPIs.
Currently, there is no provision for tracking multiple IPIs sent
to a single hart at the same time by different harts.

Use a circular queue manage the outstanding requests. Once, the
queue is full, busy wait until the there is space available in queue.
This is not the most elegant approach. It should be changed in favor
of a wakeup event once available in opensbi.

Signed-off-by: Atish Patra <atish.patra@wdc.com>