makefile: add support for building on macOS
On macOS the readlink command does not include a -f flag. Instead default
to using GNU readlink (which is often installed as greadlink).

I am not sure why readlink -f is being used here instead of $(realpath ).
If that is also an acceptable solution I'll update the PR to use that instead.

Signed-off-by: Alex Richardson <Alexander.Richardson@cl.cam.ac.uk>