Created initial set of Bazel build files required.
These include the root Bazel Workspace file, the root Bazel Build file which is required if we want to reference stuff from the root folder, as well as the Bazel Version file, which is used by Bazelisk to dowload the correct version of Bazel for the repository.

Also, added Git Ignores for the Bazel output folders. These folders contain the build artifacts produced by Bazel, so we can safely ignore them.