lib: Fix buffer overflow issue in FDT fixups
We encounter a bug in Linux kernel that some cores are failed to be
disabled and some cores are crashed during smpboot. And the root
cause of the issue is the SBI FDT fixups.
The FDT fixup functions should modify a device tree node with the
value bounded by its original value placeholder. That's why the
official method to make a Linux kernel OF device node unavailable is
to make its length to be 0. As strlen("") is surely less than
strlen("okay"), and strlen("disabled") breaks the runtime FDT.

This patch fixes the buffer overflow issue in the FDT fixups.

Reference: linux/drivers/of/base.c: __of_device_is_available()
Signed-off-by: Lv Zheng <zhenglv@smart-core.cn>