pmu: unmatched: work without mcountinhibit
mcountinhibit is absent on Unmatched U740, however it is possible to use
mcycle, mcpu and other counters without it.

mcountinhibit acts as start/stop counter, without it counters are free
running, but we still can zero or shadow the counters on start/stop.

Also firmware counters doesn't depend on HW PMU at all.

Signed-off-by: Nikita Shubin <n.shubin@yadro.com>