Makefile: Add build time and compiler info string
When we are doing opensbi development, we want to know the build time
and compiler info for debug purpose.
To enable this message, please add "BUILD_INFO=y" , like:

```
make BUILD_INFO=y
```

NOTE: `BUILD_INFO=y` will violate "reproducible builds".
So it's ONLY for development and debug purpose, and should NOT be used
in a product which follows "reproducible builds".

Signed-off-by: Wei Fu <wefu@redhat.com>