SBI: add a workaround for clang
When building with clang with `-Os -ffreestanding`, it seems that clang
improperly forms a call to `memset`.  With `-ffreestanding`, the C
standard as per Clause 4, the compiler cannot necessarily assume that
anything beyond:
- float.h
- iso646.h
- limits.h
- stdalign.h
- stdarg.h
- stdbool.h
- stddef.h
- stdint.h
- stdnoreturn.h
- fenv.h
- math.h
and the numeric conversion functions of stdlib.h.

This change explicitly converts the memset into a `sbi_memset`.  The
overhead here should be minimal, namely an extra jump for the memset
rather than an inline unrolled implementation.