sbi: scale the stack according to pointer size
Scale the default stack size according to the pointer size.  This is
important to accommodate the space required for stack allocation of the
device tree structures which scale according to the pointer size.

In the case that the default stack size is insufficient, ensure that
there is an escape hatch for users to override the default stack size at
compile time.