lib: sbi: Optionally use ".insn" instead of ".word"
".word" is used to emit a data word (32-bits) and recent GNU assembler
adds mapping symbol "$d" before ".word" to make sure that this is a data
word.  However, in lib/sbi/sbi_hfence.S, we use this to emit hfence.*
instructions (for assemblers that do not support hfence.*).  The mapping
symbol "$d" prevents these instructions being disassembled as code.

Instead, ".insn" is available to emit raw instructions but not all
assemblers with ".insn" support raw values.

This commit tests support of ".insn" with raw value in the Makefile
(AS_SUPPORT_INSN_RAW_DIRECTIVE) and optionally replaces ".word" with ".insn"
when ".insn" with raw value is supported.

Signed-off-by: Tsukasa OI <research_trasio@irq.a4lg.com>