lib: sbi: Optionally use ".option arch,+h"
On the debugger / disassembler, disassembling hfence.* as hfence.* requires
indication of the 'H'-extension use.

".option arch,+[EXT_NAME]" is one of the such ways and this commit
optionally inserts this assembler directive when supported.

It enables disassembling on the software which supports "$x[arch]" (mapping
symbols with ISA string) - latest GNU Binutils (will be released as 2.40).
Note that this commit alone won't make disassembler behavior correct on
GNU Binutils 2.39 because of a linker issue.

Signed-off-by: Tsukasa OI <research_trasio@irq.a4lg.com>